+++ resources/views/authentications/borrower.blade.php
@@ -1,0 +1,14 @@
+<div id="borrower-authentication-div">
+    <div class="row">
+        <div class="col-md-12">
+            Authentification d'un emprunteur
+        </div>
+    </div>
+    <div class="row text-center">
+        <div class="col-md-12">
+            <button class="btn btn-primary">
+                @include('icons/lender') Se connecter en tant qu'emprunteur @include('icons/lender')
+            </button>
+        </div>
+    </div>
+</div>