+++ resources/views/authentications/lender.blade.php
@@ -1,7 +1,7 @@
 <div id="lender-authentication-div">
     <div class="row">
         <div class="col-md-12">
-            <h5>Authentification d'un prêteur</h5>
+            Authentification d'un prêteur
         </div>
     </div>
     <div class="row text-center">