+++ resources/views/icons/lender.php
@@ -1,4 +1,1 @@
-<span class="fa-layers fa-fw menu-icon">
-    <i class="fas fa-circle" data-fa-transform="shrink-7 down-6 right-6" data-fa-mask="fas fa-users"></i>
-    <i class="fas fa-dice" data-fa-transform="shrink-9 down-7 right-7"></i>
-</span>
+<i class="fas fa-chess-knight"></i>