+++ resources/views/forms/new-borrowing.blade.php
@@ -1,6 +1,7 @@
 <form method="POST" action="/new-borrowing" id="new-borrowing-form">
     @csrf
     @include('authentications/lender')
+    @include('authentications/borrower')
     <div class="form-group" id="form-field-startDate">
         <label for="startDate">Date d'emprunt</label>
         <span class="input-group">