+++ changelog.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.8.0] - Unreleased
 
+### Added
+
+* New borrowing view
+    * Borrower authentication button
+
 ### Changed
 
 * Borrowings history view